refactor(game-service): modularize match.service into focused submodules
Split MatchService into domain modules under services/match/ while keeping
the public API unchanged, improving maintainability without altering game logic.
