+++ srcs/backend/game-service/src/services/match/match.utils.ts
@@ -1,0 +1,4 @@
+/** Players with an active socket connection (excludes disconnected slots). */
+export function getConnectedPlayers(match: MatchState): MatchPlayer[] {
+  return match.players.filter((entry) => entry.connected);
+}
